requirements.txt
openai-whisper
streamlit
spacy
googletrans==4.0.0rc1
torch
